Approximation of the derivative
This script calculates the incremental ratio of the function.
In this case we have velocity and we calculate acceleration.
The metod converges really well for etherogeneus data but fails when data is less etherogeneus.
In fact, for vega data, the graph is very full of spikes compared to the acceleration calculated on the simulated values.